+++ src/main/java/net/hydra/jojomod/event/index/PowerIndex.java
@@ -1,12 +1,12 @@
 package net.hydra.jojomod.event.index;
 
 public class PowerIndex {
-    public static final int NONE = 0;
-    public static final int ATTACK = 1;
-    public static final int SNEAK_ATTACK = 2;
-    public static final int GUARD = 3;
-    public static final int BARRAGE = 4;
-    public static final int MOVEMENT = 5;
-    public static final int SNEAK_MOVEMENT = 6;
+    public static final byte NONE = 0;
+    public static final byte ATTACK = 1;
+    public static final byte SNEAK_ATTACK = 2;
+    public static final byte GUARD = 3;
+    public static final byte BARRAGE = 4;
+    public static final byte MOVEMENT = 5;
+    public static final byte SNEAK_MOVEMENT = 6;
 
 }
